feat: 支持特效组播放并优化客户端帧级跟随 - 将 effect_groups 配置扩展为 key -> effect id 列表结构 - 新增服务端/客户端 playGroup、bindGroup、stopGroup API - 新增 GROUP 播放模式与特效组播放/绑定命令 - 使用 tickDelta 插值实体锚点位置，降低跟随割裂感 - 将轨迹采样改为渲染帧级更新，并使用时间驱动生命周期 - 让光束、法阵和高级特效动画使用连续渲染时间 - 补充真实辉光与编辑器预览相关配置链路

+++ src/client/kotlin/io/github/yuazer/magicrender/client/effect/trajectory/RibbonMeshBuilder.kt
@@ -136,8 +136,8 @@ object RibbonMeshBuilder {
         return if (vector.length() < 1.0E-5) fallback else vector.normalize()
     }
 
-    private fun ageT(ageTicks: Int, lifetimeTicks: Int): Double {
-        return if (lifetimeTicks <= 0) 1.0 else ageTicks.toDouble() / lifetimeTicks.toDouble()
+    private fun ageT(ageTicks: Double, lifetimeTicks: Int): Double {
+        return if (lifetimeTicks <= 0) 1.0 else ageTicks / lifetimeTicks.toDouble()
     }
 
     private fun normalizedIndex(index: Int, size: Int): Double {
